% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_model.R
\name{list_model}
\alias{list_model}
\title{list_model}
\usage{
list_model(project = "GSClassifier")
}
\arguments{
\item{project}{the project}
}
\value{
a list of models in the project
}
\description{
Explore models in a specific project
}
\examples{
list_model(project='GSClassifier')
list_model('GSClassifier')
}
\seealso{
\code{\link{lucky_model}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
