% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_project.R
\name{list_project}
\alias{list_project}
\title{list_project}
\usage{
list_project()
}
\value{
A list of project
}
\description{
list projects in \code{luckyModel} package
}
\examples{
list_project()
}
\seealso{
\code{\link{list_model}}; \code{\link{lucky_model}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
