% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucky_model.R
\name{lucky_model}
\alias{lucky_model}
\title{lucky_model}
\usage{
lucky_model(
  project = "GSClassifier",
  model = "PAD",
  developer = "HWB",
  version = "v20200110"
)
}
\arguments{
\item{project}{the project}

\item{model}{the series of models}

\item{developer}{the developer}

\item{version}{the model version}
}
\value{
a model
}
\description{
Model calling function
}
\examples{
PADi <- lucky_model(project = 'GSClassifier',
                    developer='HWB',
                    model = 'PAD',
                    version = 'v20200110')

ImmuneSubtype <- lucky_model(project = 'GSClassifier',
                            developer='Gibbs',
                            model = 'PanCancerImmuneSubtype',
                            version = 'v20190731')
}
\seealso{
\code{\link{list_project}};\code{\link{list_model}};
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
