% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_model.R
\name{save_model}
\alias{save_model}
\title{Save your model}
\usage{
save_model(
  x,
  model.path = ".",
  project = "GSClassifier",
  model = "PAD",
  developer = "HWB",
  version = "v20200110"
)
}
\arguments{
\item{x}{your model}

\item{model.path}{the path of the local repository of \code{luckyModel} package}

\item{project}{the project}

\item{model}{the series of models}

\item{developer}{the developer}

\item{version}{the model version}
}
\description{
Save your model
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
